+++ src/shared/http.interfaces.ts
@@ -26,7 +26,7 @@ export interface QueryParams {
 
 export interface ProductInterface {
   code: string,
-  status?: number;
+  status?: string;
   status_verbose?: string;
   product: {
     nutriments?: NutrimentInterface;