+++ src/components/componentDayTip/CHANGELOG.md
@@ -1,5 +1,11 @@
 # component-day-tip
 
+## 0.0.2
+
+### Patch Changes
+
+- adding deeper messages to tip of the day.
+
 ## 0.0.1
 
 ### Patch Changes